Agent,Join Req,Cancel Req,Agent Join Query,Join Req Join Query,Location Join Query

+++ src/main/java/com/alamin/toursystem/service/LocationService.java
@@ -1,8 +1,10 @@
 package com.alamin.toursystem.service;
 
 import com.alamin.toursystem.dao.LocationDao;
+import com.alamin.toursystem.dao.LocationReviewDao;
+import com.alamin.toursystem.dao.UserDao;
 import com.alamin.toursystem.entity.Location;
-import com.alamin.toursystem.entity.LocationReview;
+import com.alamin.toursystem.model.LocationReviewModel;
 import com.alamin.toursystem.exception.ResourceAlreadyExistException;
 import com.alamin.toursystem.exception.ResourceNotFoundException;
 import com.alamin.toursystem.model.LocationModel;
@@ -16,11 +18,15 @@ import java.util.List;
 public class LocationService implements LocationDao {
     @Autowired
     private LocationRepository repository;
+    @Autowired
+    private LocationReviewDao reviewDao;
     @Override
     public List<Location> getAll() {
-        List<Location> locationList=new ArrayList<>();
+
+       List<Location> locationList=new ArrayList<>();
         repository.findAll().forEach(locationList::add);
         return locationList;
+
     }
 
     @Override
@@ -30,36 +36,32 @@ public class LocationService implements LocationDao {
     }
 
     @Override
-    public Location create(LocationModel model) throws ResourceAlreadyExistException {
-        Location location=new Location(model.getLocation_name());
+    public Location create(Location model) throws ResourceAlreadyExistException {
+
         if (repository.existsById(model.getLocation_id())){
             throw new ResourceAlreadyExistException();
         }
         else {
-            Location savedName=repository.save(location);
+            Location savedName=repository.save(model);
             return savedName;
         }
     }
 
     @Override
-    public Location update(LocationModel model) throws ResourceNotFoundException {
-        /*Location location=new Location(model.getLocation_id(),
-                model.getLocation_name());
+    public Location update(Location model) throws ResourceNotFoundException {
+
         if (repository.existsById(model.getLocation_id())){
-            Location updatedLocation=repository.save(location);
+            Location updatedLocation=repository.save(model);
             return updatedLocation;
         }
         else {
             throw  new ResourceNotFoundException();
         }
-
-         */
-        return null;
     }
 
     @Override
     public Location deleteById(long location_id) throws ResourceNotFoundException {
-        /*if (repository.existsById(location_id)){
+        if (repository.existsById(location_id)){
             Location deletedLocation=repository.findById(location_id).get();
             repository.deleteById(location_id);
             return deletedLocation;
@@ -67,8 +69,35 @@ public class LocationService implements LocationDao {
         else{
             throw new ResourceNotFoundException();
         }
-
-         */
-        return null;
     }
+    @Override
+    public List<LocationModel> getAllLocation() {
+        List<Location> locations=getAll();
+        List<LocationModel> locationList=new ArrayList<>();
+
+        for (Location location:locations) {
+
+            List<LocationReviewModel> reviewList=reviewDao.findByLocationId(location.getLocation_id());
+            locationList.add(new LocationModel(
+                    location.getLocation_id(),
+                    location.getLocation_name(),
+                    reviewList
+            ));
+        }
+        return locationList;
+    }
+    @Override
+    public LocationModel findByLocationId(long location_id) throws ResourceNotFoundException {
+
+        Location location=repository.findById(location_id).orElseThrow(ResourceNotFoundException::new);
+        List<LocationReviewModel> reviewList=reviewDao.findByLocationId(location.getLocation_id());
+        LocationModel locationModel=new LocationModel(
+                location.getLocation_id(),
+                location.getLocation_name(),
+                reviewList
+        );
+        return locationModel;
+    }
+
+
 }
